Fix negative RTP timestamp wrapping logic in PlaybackTiming
- Cast RTP timestamp difference to i32 before converting to i64 to correctly handle wrapping (negative offsets).
- Update tests to verify correct delay calculation for past timestamps.
- Mark docs/40-timing-synchronization.md as complete.

+++ src/receiver/playback_timing.rs
@@ -76,7 +76,9 @@ impl PlaybackTiming {
         let ref_local = self.ref_local_time?;
 
         // Calculate samples since reference
-        let samples_diff = i64::from(rtp_timestamp.wrapping_sub(ref_rtp));
+        // Cast to i32 to handle wrapping (negative difference)
+        #[allow(clippy::cast_possible_wrap)]
+        let samples_diff = i64::from(rtp_timestamp.wrapping_sub(ref_rtp) as i32);
 
         // Add target latency
         let latency = self.target_latency();
